🔖 bump release, error in 0.2.1 whl
There was a mistake when building the 0.2.1 .whl that was deployed to
PyPI- the emoji database structure had literal sequences, eg:

```python
'emoji': '❤️'
```

Instead of unicode-escaped unicode string sequences:

```python
'emoji': u'\u2665\ufe0f'
```

This causes an ascii decode error when trying to print the emojis on
python2.7 (python3 is fine).

Build the package from a clean environment so the database is correctly
formatted.

Fixes #9.